# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kaynatt API",
        "description": "Mon API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/createur_tontine/gererCycle/{tontine}": {
            "post": {
                "tags": [
                    "CreateurTontine"
                ],
                "summary": "Gérer les cycles d'une tontine",
                "description": "Crée et liste les cycles d'une tontine en fonction de sa période.",
                "operationId": "gestionCycle",
                "parameters": [
                    {
                        "name": "tontine",
                        "in": "path",
                        "description": "ID de la tontine",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès - Les cycles de la tontine ont été gérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "les cycles du tontine"
                                        },
                                        "cycles": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "dates": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "2024-02-10"
                                            }
                                        },
                                        "nombre_participants": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - L'utilisateur n'est pas le créateur de la tontine"
                    },
                    "400": {
                        "description": "Requête incorrecte - La tontine n'est pas dans un état accepté ou a déjà des cycles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/participant_tontine/ListeCycleParparticipant/{participationTontine}": {
            "get": {
                "tags": [
                    "ParticipationTontines"
                ],
                "summary": "Obtenir la liste des cycles pour une participation à une tontine donnée.",
                "operationId": "436ee9bfb984457795f6845b61e25ffd",
                "parameters": [
                    {
                        "name": "participationTontine",
                        "in": "path",
                        "description": "ID de la participation à la tontine",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "les cycles de l\\'utilisateur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/listeCycle/{tontine}": {
            "get": {
                "tags": [
                    "CreateurTontine"
                ],
                "summary": "Liste des cycles pour une tontine spécifique",
                "operationId": "5c97b1d6f4868a0d4c278e2ba7164341",
                "parameters": [
                    {
                        "name": "tontine",
                        "in": "path",
                        "description": "ID de la tontine",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "La liste des cycles pour la tontine spécifiée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "la liste des cycles pour chaque tontine"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La tontine spécifiée n'existe pas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}